Add files via upload
The experimental and simulation data sets used in the arXiv preprint: https://arxiv.org/abs/2309.04224. the name of each file is the same as the name of the figure it is representing. The name of the files in each mat file is the same as the name of the variable plotted in the figure.